third_party/blink/web_tests/external/wpt/editing/run/justifycenter_6001-last-expected.txt

This is a testharness.js-based test.
Found 40 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div style=\\"text-align:center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div style=\\"text-align:center\\"><p>baz</p></div>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<div style=text-align:center><p>foo</div> <p>[bar]</p> <div style=text-align:center><p>baz</div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div align=\\"center\\"><p>baz</p></div>"
[FAIL] [["defaultparagraphseparator","div"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"center\\"><p>foo</p> <p>bar</p> <p>baz</p></div>" but got "<div align=\\"center\\"><p>foo</p></div> <p style=\\"text-align:center\\">bar</p> <div align=\\"center\\"><p>baz</p></div>"
[FAIL] [["defaultparagraphseparator","p"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["justifycenter",""]] "<div align=center><p>foo</div> <p>[bar]</p> <div align=center><p>baz</div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><p>foo</p></center><div style=\\"text-align:center\\"> <p>bar</p> </div><center><p>baz</p></center>" but got "<center><p>foo</p></center> <p style=\\"text-align:center\\">bar</p> <center><p>baz</p></center>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>foo</center> <p>[bar]</p> <center><p>baz</center>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false align=center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [] "<div contenteditable=false style=text-align:center><p contenteditable>f[]oo</p></div>" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "center" but got "false"
Harness: the test ran to completion.
